Persist schema tool window config to app data
