+++ src/module/query/schema-tree-node/schema-tree-node.component.scss
@@ -9,4 +9,5 @@
 
 .root-node {
     font-weight: typography.$bold;
+    text-transform: capitalize;
 }
