+++ src/service/query-tool-state.service.ts
@@ -6,8 +6,7 @@
 
 import { Injectable } from "@angular/core";
 import { FormControl } from "@angular/forms";
-import MultiGraph from "graphology";
-import { BehaviorSubject, combineLatest, first, map, Observable, ReplaySubject, shareReplay, startWith } from "rxjs";
+import { BehaviorSubject, combineLatest, first, map, Observable, shareReplay, startWith } from "rxjs";
 import { DriverAction } from "../concept/action";
 import { createSigmaRenderer, GraphVisualiser } from "../framework/graph-visualiser";
 import { defaultSigmaSettings } from "../framework/graph-visualiser/defaults";
@@ -17,7 +16,6 @@ import { INTERNAL_ERROR } from "../framework/util/strings";
 import { DriverState } from "./driver-state.service";
 import { SchemaState, Schema, SchemaAttribute, SchemaRole, SchemaConcept } from "./schema-state.service";
 import { SnackbarService } from "./snackbar.service";
-import { MatTreeFlatDataSource, MatTreeFlattener } from "@angular/material/tree";
 import {
     ApiResponse, Concept, ConceptDocument, ConceptRow, isApiErrorResponse, QueryResponse
 } from "@samuel-butcher-typedb/typedb-http-driver";
@@ -38,7 +36,6 @@ export class QueryToolState {
     queryControl = new FormControl("", {nonNullable: true});
     outputTypeControl = new FormControl("log" as OutputType, { nonNullable: true });
     outputTypes: OutputType[] = ["log", "table", "graph", "raw"];
-    readonly schemaWindow = new SchemaWindowState(this.schema);
     readonly logOutput = new LogOutputState();
     readonly tableOutput = new TableOutputState();
     readonly graphOutput = new GraphOutputState();
@@ -133,195 +130,6 @@ export class QueryToolState {
         this.tableOutput.status = "answerOutputDisabled";
         this.graphOutput.status = "answerOutputDisabled";
         this.rawOutput.push(SUCCESS_RAW);
-    }
-}
-
-export type SchemaTreeNodeKind = "root" | "concept" | "link";
-
-export interface SchemaTreeNodeBase {
-    nodeKind: SchemaTreeNodeKind;
-    children?: SchemaTreeNode[];
-    visible: boolean;
-}
-
-export interface SchemaTreeRootNode extends SchemaTreeNodeBase {
-    nodeKind: "root";
-    label: "Entities" | "Relations" | "Attributes";
-    children: SchemaTreeConceptNode[];
-}
-
-export interface SchemaTreeConceptNode extends SchemaTreeNodeBase {
-    nodeKind: "concept";
-    concept: SchemaConcept;
-    children: SchemaTreeChildNode[];
-}
-
-export type SchemaTreeLinkKind = "sub" | "owns" | "plays" | "relates";
-
-export interface SchemaTreeLinkNodeBase extends SchemaTreeNodeBase {
-    nodeKind: "link";
-    linkKind: SchemaTreeLinkKind;
-}
-
-export interface SchemaTreeSubLinkNode extends SchemaTreeLinkNodeBase {
-    linkKind: "sub";
-    supertype: SchemaConcept;
-}
-
-export interface SchemaTreeOwnsLinkNode extends SchemaTreeLinkNodeBase {
-    linkKind: "owns";
-    ownedAttribute: SchemaAttribute;
-}
-
-export interface SchemaTreePlaysLinkNode extends SchemaTreeLinkNodeBase {
-    linkKind: "plays";
-    role: SchemaRole;
-}
-
-export interface SchemaTreeRelatesLinkNode extends SchemaTreeLinkNodeBase {
-    linkKind: "relates";
-    role: SchemaRole;
-}
-
-export type SchemaTreeLinkNode = SchemaTreeSubLinkNode | SchemaTreeOwnsLinkNode | SchemaTreePlaysLinkNode | SchemaTreeRelatesLinkNode;
-
-export type SchemaTreeNode = SchemaTreeRootNode | SchemaTreeConceptNode | SchemaTreeLinkNode;
-
-export type SchemaTreeChildNode = SchemaTreeConceptNode | SchemaTreeLinkNode;
-
-export class SchemaWindowState {
-    dataSource$ = new BehaviorSubject<SchemaTreeRootNode[]>([]);
-    viewMode$ = new BehaviorSubject<"flat" | "hierarchical">("hierarchical");
-    linksVisibility$ = new BehaviorSubject<Record<SchemaTreeLinkKind, boolean>>({
-        "sub": true,
-        "owns": true,
-        "plays": true,
-        "relates": true,
-    });
-    rootNodesCollapsed: Record<SchemaTreeRootNode["label"], boolean> = {
-        "Entities": false,
-        "Relations": false,
-        "Attributes": false,
-    };
-
-    constructor(public schemaState: SchemaState) {
-        schemaState.value$.subscribe(() => {
-            this.buildView();
-        });
-        this.viewMode$.subscribe(() => {
-            this.buildView();
-        });
-        this.linksVisibility$.subscribe(() => {
-            this.updateViewVisibility();
-        });
-    }
-
-    hasChild = (_: number, node: SchemaTreeNode) => !!node.children?.length;
-
-    childrenAccessor = (node: SchemaTreeNode): SchemaTreeNode[] => {
-        return (node.children || []).filter(child => child.visible === true);
-    };
-
-    private buildView() {
-        const schema = this.schemaState.value$.value;
-        const linksVisibility = this.linksVisibility$.value;
-
-        if (!schema) {
-            this.dataSource$.next([]);
-            return;
-        }
-
-        const data: SchemaTreeRootNode[] = [{
-            nodeKind: "root",
-            label: "Entities",
-            visible: true,
-            children: Object.values(schema.entities).sort((a, b) => a.label.localeCompare(b.label)).map(x => ({
-                nodeKind: "concept",
-                concept: x,
-                visible: true,
-                children: ([
-                    ...(x.supertype ? [{ nodeKind: "link", linkKind: "sub", supertype: x.supertype, visible: linksVisibility.sub }] : []),
-                    ...x.ownedAttributes.map(y => ({ nodeKind: "link", linkKind: "owns", ownedAttribute: y, visible: linksVisibility.owns })),
-                    ...x.playedRoles.map(y => ({ nodeKind: "link", linkKind: "plays", role: y, visible: linksVisibility.plays })),
-                ] as SchemaTreeChildNode[]),
-            })),
-        }, {
-            nodeKind: "root",
-            label: "Relations",
-            visible: true,
-            children: Object.values(schema.relations).sort((a, b) => a.label.localeCompare(b.label)).map(x => ({
-                nodeKind: "concept",
-                concept: x,
-                visible: true,
-                children: ([
-                    ...(x.supertype ? [{ nodeKind: "link", linkKind: "sub", supertype: x.supertype, visible: linksVisibility.sub }] : []),
-                    ...x.relatedRoles.map(y => ({ nodeKind: "link", linkKind: "relates", role: y, visible: linksVisibility.relates })),
-                    ...x.ownedAttributes.map(y => ({ nodeKind: "link", linkKind: "owns", ownedAttribute: y, visible: linksVisibility.owns })),
-                    ...x.playedRoles.map(y => ({ nodeKind: "link", linkKind: "plays", role: y, visible: linksVisibility.plays })),
-                ] as SchemaTreeChildNode[]),
-            })),
-        }, {
-            nodeKind: "root",
-            label: "Attributes",
-            visible: true,
-            children: Object.values(schema.attributes).sort((a, b) => a.label.localeCompare(b.label)).map(x => ({
-                nodeKind: "concept",
-                concept: x,
-                visible: true,
-                children: ([
-                    ...(x.supertype ? [{ nodeKind: "link", linkKind: "sub", supertype: x.supertype, visible: linksVisibility.sub }] : []),
-                ] as SchemaTreeChildNode[]),
-            })),
-        }];
-
-        if (this.viewMode$.value === "hierarchical") {
-            const nodeMap = Object.fromEntries(data.flatMap(rootNode => rootNode.children.map(conceptNode => ([
-                conceptNode.concept.label, conceptNode
-            ]))));
-
-            data.forEach(rootNode => {
-                rootNode.children.forEach(conceptNode => {
-                    const subNode = conceptNode.children.find(node => node.nodeKind === "link" && node.linkKind === "sub");
-                    if (subNode) {
-                        const superNode = nodeMap[subNode.supertype.label];
-                        if (!superNode) throw new Error(`Missing supertype node for ${subNode.supertype.label} (nodeMap is: ${JSON.stringify(nodeMap)})`);
-                        superNode.children.unshift(conceptNode);
-                    }
-                });
-            });
-
-            data.flatMap(rootNode => rootNode.children).forEach(conceptNode => {
-                conceptNode.children = conceptNode.children.filter(child => child.nodeKind !== "link" || child.linkKind !== "sub");
-            });
-
-            data.forEach(x => x.children = x.children.filter(conceptNode => !conceptNode.concept.supertype));
-        }
-
-        this.dataSource$.next(data);
-    }
-
-    private updateViewVisibility() {
-        this.dataSource$.value.forEach(x => x.children.forEach(conceptNode => {
-            conceptNode.children.filter(child => child.nodeKind === "link").forEach(linkNode => {
-                linkNode.visible = this.linksVisibility$.value[linkNode.linkKind];
-            });
-        }));
-        this.dataSource$.next([...this.dataSource$.value]);
-    }
-
-    useFlatView() {
-        this.viewMode$.next("flat");
-    }
-
-    useHierarchicalView() {
-        this.viewMode$.next("hierarchical");
-    }
-
-    toggleLinksVisibility(linkKind: SchemaTreeLinkKind) {
-        this.linksVisibility$.next({
-            ...this.linksVisibility$.value,
-            [linkKind]: !this.linksVisibility$.value[linkKind]
-        });
     }
 }
